+++ tests/test_avalia_nota.py
@@ -6,7 +6,7 @@ from ..src import nlp_predictor
 @pytest.fixture
 def nlp_object():
     '''Cria um objeto que será utilizado pelo pytest'''
-    return nlp_predictor(retrain_model=False)
+    return nlp_predictor(model_type = 'BERT', retrain_model=False)
 
 
 def test_avalia_nota(nlp_object) -> None: